Add true cache miss detection
WordPress' `get_transient()` returns `false` if transient is not found. `false` is also a valid value, however, and therefore `get()` may get false negatives, in which case the default value is returned. This is sub-optimal.

Now, if `get_transient()` returns `false`, an additional check is performed, looking up the transient option using a more low-level `get_option()` method. The class' default value is used to determine if the option exists or not. WP's object cache means that a query will often not be made, and the lookup only happens to confirm false negatives, and not when result is positive, meaning that in the majority of cases the additional check will not happen. Of course, `has()` now uses a similar approach.